REFACTOR exit status is not a uint8_t

+++ include/error.h
@@ -6,7 +6,7 @@
 /*   By: lbrusa <lbrusa@student.42berlin.de>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/19 11:23:18 by lbrusa            #+#    #+#             */
-/*   Updated: 2024/06/09 11:26:30 by lbrusa           ###   ########.fr       */
+/*   Updated: 2024/06/09 14:30:49 by lbrusa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@ extern "C" {
 
 # include "minishell.h"
 
-int		status_and_perror(const char *msg, int status);
-int     status_and_detailed_perror(const char *msg_1, const char *msg_2, int status);
-int		exit_and_print_err(const char *msg, int status);
-int     print_error_status(const char *message, int status);
-int	    print_minishell_error_status(const char *message, int status);
-int		zero_and_printerr(const char *msg);
+uint8_t	status_and_perror(const char *msg, uint8_t status);
+uint8_t	status_and_detailed_perror(const char *msg_1, const char *msg_2, uint8_t status);
+uint8_t	exit_and_print_err(const char *msg, uint8_t status);
+uint8_t	print_error_status(const char *message, uint8_t status);
+uint8_t	print_minishell_error_status(const char *message, uint8_t status);
+uint8_t	zero_and_printerr(const char *msg);
 void	*null_on_err(const char *message);
 bool	false_and_perr(const char *msg);
 bool	false_and_print(const char *msg);
